AssetBrowser ImGui - Change fb_width/fb_height to floats.

+++ apps/asset_browser/src/renderer.cpp
@@ -174,8 +174,8 @@ namespace mill::asset_browser
     void Renderer::render_draw_data(u64 context_id, const ImDrawData* draw_data)
     {
         // Avoid rendering when minimized, scale coordinates for retina displays (screen coordinates != framebuffer coordinates)
-        const u32 fb_width = (int)(draw_data->DisplaySize.x * draw_data->FramebufferScale.x);
-        const u32 fb_height = (int)(draw_data->DisplaySize.y * draw_data->FramebufferScale.y);
+        const f32 fb_width = draw_data->DisplaySize.x * draw_data->FramebufferScale.x;
+        const f32 fb_height = draw_data->DisplaySize.y * draw_data->FramebufferScale.y;
         if (fb_width <= 0 || fb_height <= 0)
             return;
